update hook for show toast token
